backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
supabase>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic[email]>=2.5.0,<3.0.0